==> 2_selection_introgression/genome_features/04_process_positive_selection.ipynb <==
## Process RAiSD output for downstream analysis

* Read in test data, open a new output file, then loop through each line
* If the line matches the scaffold heading pattern (starts with '/'), check if there is SNP data on the next line
* If SNP data, save scaffold to variable to be pasted to each subsequent line, until hitting the next scaffold heading

### Start with test dataset

* 1000 lines and 4 scaffolds (mtDNA + 3 nuclear)
* first part of coastal output

file1=open("test.coastal")

line1=file1.readlines()
file2=open("output.test.coastal","w")

end=len(line1)

for i in range(0,end):

    if (line1[i][0]!='/'):
        newLine=scaffold.strip('\n')+'\t'+line1[i]
        file2.write(newLine)
    elif (line1[i][0]=='/')&(line1[i+1][0]=='/'):
        print('no data for scaffold: '+line1[i][3:])
    elif (line1[i][0]=='/')&(line1[i+1][0]!='/'):
        scaffold=line1[i][3:]
        print('start scaffold: '+scaffold)
    else:
        print("problem: line "+str(i))

file2.close()
file1.close()
    

### Process coastal RAiSD output

file1=open("RAiSD_Report.coastal")

line1=file1.readlines()
file2=open("RAiSD.output.coastal","w")
end=len(line1)

for i in range(0,end):
    
    if (line1[i][0]!='/'):
        newLine=scaffold.strip('\n')+'\t'+line1[i]
        file2.write(newLine)
    elif (line1[i][0]=='/')&(line1[i+1][0]=='/'):
        print('no data for scaffold: '+line1[i][3:])
    elif (line1[i][0]=='/')&(line1[i+1][0]!='/'):
        scaffold=line1[i][3:]
        print('start scaffold: '+scaffold)
    else:
        print("problem: line "+str(i))

file2.close()
file1.close()
    


### Process inland RAiSD output

file1=open("RAiSD_Report.inland")

line1=file1.readlines()
file2=open("RAiSD.output.inland","w")
end=len(line1)

for i in range(0,end):
    
    if (line1[i][0]!='/'):
        newLine=scaffold.strip('\n')+'\t'+line1[i]
        file2.write(newLine)
    elif (line1[i][0]=='/')&(line1[i+1][0]=='/'):
        print('no data for scaffold: '+line1[i][3:])
    elif (line1[i][0]=='/')&(line1[i+1][0]!='/'):
        scaffold=line1[i][3:]
        print('start scaffold: '+scaffold)
    else:
        print("problem: line "+str(i))

file2.close()
file1.close()
